server: add support for specifying content type in rmeta call (#5)
* add support for specifying content type in rmeta call

This commit is intended to allow users of this library to
leverage the full capabilities of the Tika APIs. In particular,
the API for the /rmeta endpoint allows for options other than
the currently hard-coded "text" option [1]. This change adds
a method to the API to facilitate specifying other types,
while leaving the functionality of the existing method
unchanged.

[1] https://wiki.apache.org/tika/TikaJAXRS#Recursive_Metadata_and_Content

* add tests for MetaRecursiveType

Following on on change request from PR: https://github.com/google/go-tika/pull/5#pullrequestreview-170014465

The style here is a bit different from some of the other
tests because it seemed to make sense to emulate the behavior
of the tika server a bit more closely for this use case.