support passing http.Header to exported functions
In doing this we allow the user of the library to pass nil, getting the current default behaviour, or
specify a header. This header can then be used to indicate to the Tika server the response format that is required.

This should be useful to folks because Tika can respond with HTML, plain text, or json. At this point I've not added
any validation in here. My assumption is that the Tika server will handle that.

I updated the tests to nimply use nil for this new parameter. If it is wanted I could look to add some new tests that
exercies different response types. It may also be nice to update the cli tool to allow the user to indicate the
type of response required. Given the definition of the issue though that felt out of scope here.

I hope this is helpful.