introduce WithHeader functions
This adds two new Parse functions, ParseWithHeader and ParseReaderWithHeader.
I've also added a specific test case here for the ParseWithHeader function to ensure it
is using the header correctly.

The other change here, which no longer breaks the API, is to have the original Parse and
ParseReader simply call the new WithHeader versions setting a nil header.
This approach means that we still only have one place where call and callString are called
but that we also have the ability to have headers or not.